DHT-RPC request ids are only u16 numbers
The protobuf encoding uses u64 space, but the
[implementation is only](https://github.com/mafintosh/dht-rpc/blob/19a0df46cd63ec0ddfff16b7ffab77d384e0c4db/lib/io.js#L49)
u16.

This was causing interop issues and generating `InResponseBadRequestId`
as the number we receive was not being found on the request/response
map.

+++ src/rpc/io.rs
@@ -220,7 +220,7 @@ where
 
     fn request(&mut self, mut ev: MessageEvent<TUserData>) {
         let (msg, peer) = ev.inner_mut();
-        msg.rid = self.next_req_id().0;
+        msg.rid = self.next_req_id().0 as u64;
 
         if msg.is_holepunch() && peer.referrer.is_some() {
             match &ev {
